Refactored: put create_job into business_job_tasks
